// src/interface/course.interface.tsx
export interface CourseInterface{
    	courseId: string;
		courseName: string;
		courseDescription: string; 
		facultyDepartment: string;
		academicTerm: string;
		academicYear: Number; 
		professors: string;
		prerequisites: string; 
		status: string;
		curriculumName: string;
		degreeLevel: string;
		teachingHours: Number;  
}